backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyserial==3.5
hidapi==0.14.0
pyusb==1.2.1
numpy==1.24.3
requests==2.31.0
pydantic==2.4.2
python-multipart==0.0.6
influxdb-client==1.36.1
